Update FilterPaneTagLib.groovy
Fixed problem that caused IllegalArgumentException while enum value was not set. The exception was caught silently so it was not obvious that it was thrown. I saw it while debugging.

Also fixed problem that valueMessagePrefix was not used for enum values unless values were explicitly specified. Removing the optionValue from ctrlAttrs made the trick.