Solution for https://jira.grails.org/browse/GPFILTERPANE-131
See the JIRA description for details.

I've lower-cased the case values so that they will match against the lower-cased filter operation.

I've also specified that if the filterValue is a date field ('date.struct' as a value), it will attempt to parse it. Previously, the check was only for 'struct', and I've left this in in case there was some behaviour being relied upon that used that value (though given that the case statements were not evaluating this would be surprising).